+++ CHANGELOG.md
@@ -16,6 +16,9 @@
  * Android will not warn for usage of deprecated `useJack` option anymore.
  * Upgraded Checkstyle to 7.1
  * PMD can now cache remote config files, and use them when offline.
+ * Android lint XML report is now written under `$buildDir/reports/android/android-lint.xml`
+    for consistency with other tools. This name is the same regardless of the used plugin version.
+    HTML report is untouched.
 
 # v2.1
 * **2.1.8**